+++ fr/main/model/Universe.java
@@ -1,6 +1,5 @@
 package fr.main.model;
 
-import java.io.*;
 import java.awt.Dimension;
 import java.awt.Point;
 import java.io.File;