+++ fr/main/model/units/air/Fighter.java
@@ -9,6 +9,7 @@ import fr.main.model.units.AbstractUnit;
 import fr.main.model.units.MoveType;
 import fr.main.model.units.Unit;
 import fr.main.model.units.weapons.PrimaryWeapon;
+import fr.main.view.sound.MusicEngine;
 
 /**
  * Represents a fighter
@@ -36,10 +37,18 @@ public class Fighter extends Unit implements PlaneUnit{
 
     public Fighter(Player p, int x, int y){
         this(p,new Point(x,y));
+        initialSound();
     }
+    
+    public void initialSound() {
+		this.selected = new MusicEngine("./assets/sound/song021.wav");
+		this.attack = new MusicEngine("./assets/sound/song037.wav");
+		//this.died = new MusicEngine(null);
+	}
 
     public Fighter(Player player, Point point){
         super(player,point,fuelName,99,true,MoveType.AIRY,9,2,new PrimaryWeapon(PRIMARYWEAPON_NAME,9,PRIMARYWEAPON_DAMAGES,true),null,NAME,PRICE);
+        initialSound();
     }
 
     public int getFuelTurnCost(){