+++ fr/main/model/units/land/Infantry.java
@@ -11,6 +11,7 @@ import fr.main.model.units.AbstractUnit;
 import fr.main.model.units.Unit;
 import fr.main.model.units.MoveType;
 import fr.main.model.units.weapons.SecondaryWeapon;
+import fr.main.view.sound.MusicEngine;
 import fr.main.model.units.CaptureBuilding;
 import fr.main.model.units.air.BCopter;
 import fr.main.model.units.air.TCopter;
@@ -48,16 +49,24 @@ public class Infantry extends Unit implements WalkingUnit,CaptureBuilding{
         SECONDARYWEAPON_DAMAGES.put(APC.class,14);
         SECONDARYWEAPON_DAMAGES.put(BCopter.class,7);
         SECONDARYWEAPON_DAMAGES.put(TCopter.class,30);
-
+        
         SECONDARYWEAPON = new SecondaryWeapon(SECONDARYWEAPON_NAME,SECONDARYWEAPON_DAMAGES);
     }
 
 	public Infantry(Player p, int x, int y){
 		this(p,new Point(x,y));
+		initialSound();
+	}
+	
+	public void initialSound() {
+		this.selected = new MusicEngine("./assets/sound/song069.wav");
+		this.attack = new MusicEngine("./assets/sound/ak47plus.wav");
+		//this.died = new MusicEngine(null);
 	}
 
 	public Infantry(Player player, Point point){
 		super(player,point,WalkingUnit.FUEL_NAME,99,false,MoveType.INFANTRY,3,2,null,SECONDARYWEAPON,NAME,PRICE);
+		initialSound();
 	}
 
     public boolean canCapture(AbstractBuilding b){