+++ fr/main/view/controllers/GameController.java
@@ -1,26 +1,49 @@
 package fr.main.view.controllers;
 
-import java.awt.event.*;
-import java.awt.*;
-import javax.swing.SwingUtilities;
-import java.util.*;
+import java.awt.Dimension;
+import java.awt.Graphics;
+import java.awt.Image;
+import java.awt.Point;
+import java.awt.event.KeyEvent;
+import java.awt.event.MouseEvent;
+import java.awt.event.MouseWheelEvent;
 import java.io.File;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.LinkedList;
+import java.util.Map;
+import java.util.Scanner;
 
+import javax.swing.SwingUtilities;
+
+import fr.main.model.Direction;
+import fr.main.model.Universe;
+import fr.main.model.buildings.AbstractBuilding;
+import fr.main.model.buildings.FactoryBuilding;
+import fr.main.model.buildings.MissileLauncher;
+import fr.main.model.buildings.OwnableBuilding;
 import fr.main.model.players.Player;
-import fr.main.model.*;
-import fr.main.model.units.*;
-import fr.main.model.units.naval.NavalUnit;
-import fr.main.model.buildings.*;
-
-import fr.main.view.*;
-import fr.main.view.views.GameView;
-import fr.main.view.interfaces.*;
+import fr.main.model.units.AbstractUnit;
+import fr.main.model.units.CaptureBuilding;
+import fr.main.model.units.HealerUnit;
+import fr.main.model.units.HideableUnit;
+import fr.main.model.units.SupplyUnit;
+import fr.main.model.units.TransportUnit;
+import fr.main.view.MainFrame;
+import fr.main.view.Position;
+import fr.main.view.interfaces.ActionPanel;
+import fr.main.view.interfaces.BuildingInterface;
+import fr.main.view.interfaces.DayPanel;
+import fr.main.view.interfaces.InterfaceUI;
+import fr.main.view.interfaces.Minimap;
+import fr.main.view.interfaces.PlayerPanel;
+import fr.main.view.interfaces.TerrainPanel;
 import fr.main.view.render.PathRenderer;
 import fr.main.view.render.UniverseRenderer;
-import fr.main.view.render.units.UnitRenderer;
 import fr.main.view.render.buildings.BuildingRenderer;
 import fr.main.view.render.sprites.Sprite;
-import fr.main.view.sound.Sdfx;
+import fr.main.view.render.units.UnitRenderer;
+import fr.main.view.views.GameView;
 
 /**
  * Game Controller.
@@ -211,6 +234,7 @@ public class GameController extends Controller {
         }
     }
 
+    // add the selectiong item in the up-right panel
     private class MainActionPanel extends ControllerPanel {
 
         public MainActionPanel () {
@@ -226,6 +250,7 @@ public class GameController extends Controller {
             new Index("Big power", world::bigPower);
             new Index("Small power", world::smallPower);
             new Index("Save", world::save);
+            new Index("Options", () -> MainFrame.setScene(new OptionController(GameController.this)));
             new Index("Quit to menu", () -> MainFrame.setScene(new MenuController()));
             new Index("Quit game", () -> System.exit(0));
         }
@@ -241,7 +266,8 @@ public class GameController extends Controller {
 
     private class TransportUnitPanel extends ControllerPanel {
 
-        private TransportUnit transportUnit;
+        @SuppressWarnings("unused")
+		private TransportUnit transportUnit;
         public AbstractUnit selected = null;
 
         public TransportUnitPanel (){
@@ -511,7 +537,6 @@ public class GameController extends Controller {
                 defaultValues();
             }
         }
-
         /**
          * Sets the controls to their default value
          */
@@ -569,7 +594,12 @@ public class GameController extends Controller {
                             (target.getX() - camera.getX() + 1) * MainFrame.UNIT + 5,
                             (target.getY() - camera.getY()) * MainFrame.UNIT + 5, 1000,
                             UniverseRenderer.FlashMessage.Type.ALERT);
-                Sdfx.EXPLOSION.play();
+                targetUnit.getMusicEngine("attack").start(false);
+                if(target.getLife()==0) {
+                	target.getMusicEngine("died").start(false);
+                }else if(targetUnit.getLife()==0) {
+                	targetUnit.getMusicEngine("died").start(false);
+                }
             } else {
                 targetUnit.setMoveQuantity(0);
                 if (targetUnit.getPrimaryWeapon() != null) targetUnit.getPrimaryWeapon().shoot();
@@ -595,7 +625,7 @@ public class GameController extends Controller {
             else buildingPanel.setVisible (true);
         } else if (targetUnit.getPlayer() == world.getCurrentPlayer() &&
                  targetUnit.isEnabled()) {
-            Sdfx.SELECT.play();
+        	targetUnit.getMusicEngine("selected").start(false);
             mode = Mode.UNIT;
             world.updateTarget(targetUnit);
             path.rebase(targetUnit);
@@ -675,7 +705,30 @@ public class GameController extends Controller {
      * Moves the camera and cursor to the next enabled unit
      */
     public void findNextUnit(){
-        //TODO
+        AbstractUnit unit = targetUnit;
+        ArrayList<AbstractUnit> units = world.getCurrentPlayer().unitList();
+
+        int n = (units.indexOf(unit) + units.size()) % units.size(),
+            // indexOf returns -1 if the object is not in the list, here we want to have the last element in this case
+            i = (n + 1) % units.size();
+            // the first element we check
+        unit = null;
+        do {
+            if (units.get(i).isEnabled()){
+                unit = units.get(i);
+                break;
+            }
+            i = (i + 1) % units.size();
+        } while (i != n);
+
+        if (unit == null && units.get(n).isEnabled())
+            unit = units.get(n);
+
+        if (unit != null){
+            camera.setLocation(unit.getX(), unit.getY());
+            cursor.setLocation(unit.getX(), unit.getY());
+            unitCursor.setLocation(unit.getX(), unit.getY());
+        }
     }
 
     /**