+++ fr/main/view/views/MenuView.java
@@ -1,9 +1,11 @@
 package fr.main.view.views;
 
-import java.io.IOException;
-import javax.swing.*;
-import fr.main.view.components.*;
-import fr.main.view.controllers.*;
+import javax.swing.ImageIcon;
+import javax.swing.JButton;
+import javax.swing.JLabel;
+
+import fr.main.view.components.MenuButton;
+import fr.main.view.controllers.MenuController;
 
 /**
  * Rendering main menu
@@ -29,24 +31,20 @@ public class MenuView extends View {
     
     public MenuView (MenuController controller) {
         super(controller);
-        // TODO Auto-generated constructor stub
-        try {
-            bg     = new ImageIcon("./assets/screens/bd.png");
-            label  = new JLabel(bg);
-            
-            play   = new MenuButton("PLAY","./assets/button/b02.png", 350, 150);
-            load   = new MenuButton("LOAD","./assets/button/b02.png", 350, 290);
-            edit   = new MenuButton("MAP EDITOR","./assets/button/b02.png", 350, 430);
-            host   = new MenuButton("HOST","./assets/button/b02.png", 350, 430);
-            join   = new MenuButton("JOIN","./assets/button/b02.png", 350, 570);
-            
-            exit   = new MenuButton("EXIT","./assets/button/b03.png", 830, 0, 20);
-            option = new MenuButton("OPTION","./assets/button/b03.png",-5, 0, 20);
-            sound  = new MenuButton("", "./assets/button/music02.png", 40, 600);
-            
-        } catch (IOException e) {
-            System.err.println(e);
-        }
+
+        bg     = new ImageIcon("./assets/screens/bd.png");
+		label  = new JLabel(bg);
+		
+		play   = new MenuButton("PLAY","./assets/button/b02.png", 350, 150);
+		load   = new MenuButton("LOAD","./assets/button/b02.png", 350, 290);
+		host   = new MenuButton("HOST","./assets/button/b02.png", 350, 430);
+		join   = new MenuButton("JOIN","./assets/button/b02.png", 350, 570);
+        edit   = new MenuButton("MAP EDITOR","./assets/button/b02.png", 350, 430);
+		
+		exit   = new MenuButton("EXIT","./assets/button/b03.png", 830, 0, 20);
+		option = new MenuButton("OPTION","./assets/button/b03.png",-5, 0, 20);
+		sound  = new MenuButton("", "./assets/button/music03.png", 40, 600);
+
 
         label.setBounds(0, 0, bg.getIconWidth(), bg.getIconHeight());
         label.setIcon(bg);
@@ -71,6 +69,7 @@ public class MenuView extends View {
         host.addActionListener(controller.host);
         join.addActionListener(controller.join);
         option.addActionListener(controller.option);
+        controller.bm.start(true);
         controller.new Music(sound);
     }